Do not use JUnit Suite during build; call regressiçon directly
